feat: add client role service, repository and interface

+++ internal/dto/client_role/create.go
@@ -7,6 +7,6 @@ type CreateClientRoleReq struct {
 
 type CreateClientRoleRes struct {
 	Id       int64 `json:"id"`
-	ClientId int64 `json:"clientId"`
-	RoleId   int64 `json:"roleId"`
+	ClientId int32 `json:"clientId"`
+	RoleId   int32 `json:"roleId"`
 }
