+++ internal/dto/client_role/get.go
@@ -6,6 +6,6 @@ type GetClientRoleReq struct {
 
 type GetClientRoleRes struct {
 	Id       int64 `json:"id"`
-	ClientId int64 `json:"clientId"`
-	RoleId   int64 `json:"roleId"`
+	ClientId int32 `json:"clientId"`
+	RoleId   int32 `json:"roleId"`
 }
